pyproject.toml
[project]
name = "pdf_summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "pdfminer-six>=20250506",
    "requests>=2.32.4",
]
